info: error on loading wrong BPF object type
Before this patch, ebpf.LoadPinnedMap and ebpf.LoadPinnedProg both
succeed when used with the wrong object type. Meaning that you can open
a prog using epbf.LoadPinnedMap or vice-versa and you'll get garbage
data in the object's info.

To my knowledge, there are two ways of knowing from an opened FD which
object type it is:
* checking for specific fields in /proc/self/fdinfo/<fd>;
* running readlink(2) on /proc/self/fd/<fd> and check for
  anon_inode:bpf-prog or anon_inode:bpf-map.

This is a breaking change for users relying on LoadPinned<object> to
open any object: I've been using that behavior to scan BPF filesystem.

Signed-off-by: Mahe Tardy <mahe.tardy@gmail.com>